Fixed up nav bar for entries page, updates now with new changes
